Move webscaper modules to utils Clean placeholder function
