--- dags/dag_fritidsjobs_webscraper/fritidsjobs_webscraper_data.py
@@ -1,11 +1,0 @@
-import logging
-
-
-logger = logging.getLogger(__name__)
-
-
-def placeholder_function() -> None:
-    """
-    Placeholder function for processing the fritidsjobs_webscraper data.
-    """
-    logger.info("Processing fritidsjobs_webscraper data...")
